requirements.txt
flask
requests
pytest
Werkzeug
flask-sqlalchemy
flask-login